Add "sslmode" setting for PostgreSQL connection
The commit laravel/framework@586bffa added support for sslmode in
PostgresConnector.php and sslmode has been around since postgres version
9.1 (2011).

This change makes it possible to specify sslmode from the config file.

Also serves as documentation to other developers so they don't have to
dive deep into the code to figure out that it's posible to set this
option.

The posible values for sslmode are:
    disable, allow, prefer, require, verify-ca, verify-full

The default value is "prefer".

Signed-off-by: crynobone <crynobone@gmail.com>