Wrap .env.example variable values in quotes
When the developer copies .env.example to .env they are unlikely to
add quotes to the values in .env. When the developer needs to
set the value of an environment variable to a value containing a
space, as none of the existing values in .env are quoted, the
developer is unlikely to wrap this new value with spaces in quotes.
This will result in an error, as The vlucas/phpdotenv library throws
an error when setting an environment variable to a value with a space.

Quote all default environment variables by default, reducing the
likelihood of the developer ever receiving the error relating to
environment variables needing to be quoted when they contain spaces.

This excludes true, null etc. even though wrapping them in quotes
would still result in the desired behaviour, as it is more intuitive
to see these special types not wrapped in quotes.

The current default values in .env.example don't contain any spaces,
so this commit will make no difference out of the box; its only
purpose is to help out the developer further down the line on the day
when they need to set an environment variable to a value containing
spaces.

Also, the vlucas/phpdotenv library docs currently uses quoted
variables in every example, so quoting .env.example's values will lead
to more consistency with those docs. This will likely stop any
confusion for the developer when reading the vlucas/phpdotenv docs.

Signed-off-by: crynobone <crynobone@gmail.com>

Conflicts:
	.env.example